requirements.txt
langchain-core
langchain-google-genai
langchain-community 
langchain_huggingface
pypdf
python-dotenv
colorist
chromadb